Disable wireguard for windows
